#28 WIP implementation finished
- bug: all times (creation, retention)  should be int64
- switching UnixNano() / 1000000 to Unix()Milli()
- SFMV implementation finished

+++ pkg/model/retentionentry.go
@@ -4,16 +4,16 @@ type RetentionEntry struct {
 	Filename      string `yaml:"filename" json:"filename"`
 	TenantID      string `yaml:"tenantID" json:"tenantID"`
 	BlobID        string `yaml:"blobID" json:"blobID"`
-	CreationDate  int    `yaml:"creationDate" json:"creationDate"`
+	CreationDate  int64  `yaml:"creationDate" json:"creationDate"`
 	Retention     int64  `yaml:"retention" json:"retention"`
-	RetentionBase int    `yaml:"retentionBase" json:"retentionBase"`
+	RetentionBase int64  `yaml:"retentionBase" json:"retentionBase"`
 }
 
 func (r *RetentionEntry) GetRetentionTimestampMS() int64 {
 	if r.RetentionBase > 0 {
-		return int64(r.RetentionBase) + r.Retention*60*1000
+		return r.RetentionBase + r.Retention*60*1000
 	}
-	return int64(r.CreationDate) + r.Retention*60*1000
+	return r.CreationDate + r.Retention*60*1000
 }
 
 func RetentionEntryFromBlobDescription(b BlobDescription) RetentionEntry {
